Smart uc: fixes for Ubuntu 12.04.1, 64-bit

Directly merge this in SmartUC; will need to be cleaned up before merging in experimental

+++ utils.sh
@@ -56,7 +56,11 @@ get_exec_libs() {
       then
        echo Warning! Not finding some library...
       else 
-       cp $L $2
+       if [ ! -e $2/$(basename $L) ]
+        then
+         cp $L $2
+         get_exec_libs $L $2
+        fi
       fi
    done
 }
